chore(cc-review): refactor code to pass cc review
- refactor code to pass cc review

[Finishes #159550781]
