+++ client/tests/components/shared/Form/MenuForm.spec.jsx
@@ -102,12 +102,22 @@ describe('MenuForm', () => {
   });
 
   it('changes date state on date change', () => {
+    // Date.now() =
     const wrapper = mount(<MenuForm {...props} />);
     const event = { target: { value: '2018-04-27' } };
 
     wrapper.find('input').at(0).simulate('change', event);
 
     expect(wrapper.state().date).toEqual('2018-04-27');
+  });
+
+  it('does not changes date state on date change when date is lower than current date', () => {
+    const wrapper = mount(<MenuForm {...props} />);
+    const event = { target: { value: '1969-01-01' } };
+
+    wrapper.find('input').at(0).simulate('change', event);
+
+    expect(wrapper.state().date).toEqual('1970-01-01');
   });
 
   it('submits Menu Form when adding menu', () => {
