+++ package.json
@@ -51,6 +51,9 @@
     "body-parser": "^1.18.2",
     "classnames": "^2.2.5",
     "clean-webpack-plugin": "^0.1.19",
+    "compression": "^1.7.3",
+    "compression-webpack-plugin": "^1.1.11",
+    "connect-gzip-static": "^2.1.1",
     "cors": "^2.8.4",
     "cross-env": "^5.1.4",
     "css-loader": "^0.28.11",
