+++ webpack.prod.js
@@ -4,6 +4,7 @@ const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
+const CompressionPlugin = require('compression-webpack-plugin');
 const merge = require('webpack-merge');
 const common = require('./webpack.common.js');
 
@@ -19,6 +20,14 @@ const uglifyPlugin = new UglifyJsPlugin({
   cache: true,
   parallel: true,
   sourceMap: true // set to true if you want JS source maps
+});
+
+const compressionPlugin = new CompressionPlugin({
+  asset: '[path].gz[query]',
+  algorithm: 'gzip',
+  test: /\.js$|\.css$|\.html$/,
+  threshold: 10240,
+  minRatio: 0.8
 });
 
 dotenv.config();
@@ -58,6 +67,7 @@ module.exports = merge(common, {
     envPlugin,
     cssPlugin,
     uglifyPlugin,
-    optimizeCSSPlugin
+    optimizeCSSPlugin,
+    compressionPlugin
   ],
 });
